Add files via upload
 Space Mission Launches Data Analytics Project

*by Franco Maldonado*

This is part of Space Mission Launches Data Analytics Data: Cleaning, analysis and visualization
and this workbook has been created to document the entire process of data analysis, cleaning and visualization, and to show the skills and tools to be used. 

- Project background

I Worked with an incredibly rich dataset from nextspaceflight.com that includes all space missions since the beginning of the US-Soviet space race in 1957. 
It contains data on mission status (success/failure), mission cost, number of launches per country and much more. There is much we can learn from this data set about dominant organizations and trends over time.